refactor: add custom hook useEditInputLogic
