+++ src/hooks/useEditInputLogic.ts
@@ -1,0 +1,68 @@
+import { useStoreEditForm } from "@/store/storeEditForm";
+import { getTimeFromSeconds } from "@/utils/helpers";
+
+export function useEditInputLogic(name: string) {
+    const { editedTime, setEditedTimeByValues } = useStoreEditForm();
+    const [hours, minutes, seconds] = getTimeFromSeconds(editedTime);
+
+    const getCurrentValue = () => {
+        if (name === "hours") return hours;
+        if (name === "minutes") return minutes;
+        if (name === "seconds") return seconds;
+        return 0;
+    };
+
+    const changeHandle = (e: React.ChangeEvent<HTMLInputElement>) => {
+        const value = e.target.value;
+        if (/^\d*$/.test(value)) {
+            const newValue = Number(value);
+            if (name === "hours") {
+                setEditedTimeByValues(newValue, minutes, seconds);
+            } else if (name === "minutes") {
+                setEditedTimeByValues(hours, newValue, seconds);
+            } else if (name === "seconds") {
+                setEditedTimeByValues(hours, minutes, newValue);
+            }
+        }
+    };
+
+    const wheelHandle = (e: React.WheelEvent<HTMLInputElement>, type: string) => {
+        e.preventDefault();
+        const delta = e.deltaY > 0 ? -1 : 1;
+        if (type === "hours") {
+            let changedHours = hours + delta;
+            if (changedHours > 24) {
+                changedHours = 0;
+            }
+            if (changedHours < 0) {
+                changedHours = 24;
+            }
+            setEditedTimeByValues(changedHours, minutes, seconds);
+        }
+        if (type === "minutes") {
+            let changedMinutes = minutes + delta;
+            if (changedMinutes > 59) {
+                changedMinutes = 0;
+            }
+            if (changedMinutes < 0) {
+                changedMinutes = 59;
+            }
+            setEditedTimeByValues(hours, changedMinutes, seconds);
+        }
+        if (type === "seconds") {
+            let changedseconds = seconds + delta;
+            if (changedseconds > 59) {
+                changedseconds = 0;
+            }
+            if (changedseconds < 0) {
+                changedseconds = 59;
+            }
+            setEditedTimeByValues(hours, minutes, changedseconds);
+        }
+    };
+    return {
+        value: getCurrentValue(),
+        changeHandle,
+        wheelHandle,
+    };
+}
